Fixed pulse generator
Fixed a problem with the pulse generator not working, which you could make work by going back to an earlier commit. It appears the EBSR register is sensitive, and is therefore only written to once now, and required bit settings are subsequently checked in the expansion board- and AIC3204 initialization code.